suppression de la page article detail, renommage articledetailcopy
