+++ src/App.js
@@ -7,7 +7,7 @@ import FirstVisit from './screens/FirstVisit'
 import IdentificationAdmin from './screens/IdentificationAdmin'
 import Home from './screens/Home'
 import PanelAdmin from './screens/PanelAdmin'
-import ArticleDetail from './screens/ArticleDetailcopy'
+import ArticleDetail from './screens/ArticleDetail'
 import ArticleContextProvider from './context/ArticleContext'
 import CategoriesContextProvider from './context/CategoriesContext'
 import SousCategoriesContextProvider from './context/SousCategoriesContext'
