+++ src/screens/ArticleDetail.js
@@ -5,7 +5,7 @@ import { useParams, Link } from "react-router-dom";
 import "./Styles/ArticleDetail.css";
 import fleche from "../assets/icones/flecheCercle.png";
 
-const ArticleDetail1 = () => {
+const ArticleDetail = () => {
   const { articles } = useContext(ArticleContext);
   console.log(articles);
   let { id } = useParams();
@@ -25,13 +25,13 @@ const ArticleDetail1 = () => {
               return (
                 <>
                   <div className="stateCatArticlesCat">
-                    <Link to="/articlesGrid">
-                      <button className="btnCat ">[ CATEGORIE ]</button>
+                    <Link to="/articlesGrid"> 
+                      <button className="btnCat ">{result.nom_categorie}</button>
                     </Link>
                     <img src={fleche} alt="fleche" className="gridArrow" />
                     <Link to="/articlesGrid">
                       <button className="btnSousCat ">
-                        [ SOUS-CATEGORIE ]
+                      {result.nom_sous_categorie}
                       </button>
                     </Link>
                   </div>
@@ -39,35 +39,31 @@ const ArticleDetail1 = () => {
                   <div className="articleDetailIntro">
                     <p id="articleDetailIntro">{result.intro}</p>
                   </div>
-
-                  <br />
-                  <br />
-
+                  <div className="articleDetailImage">
+                    <img
+                      src={result.image}
+                      alt={result.tire}
+                      className="articleImage"
+                    ></img>
+                  </div>
                   <div className="articleDetailFirst">
-                    <div className="articleDetailImage">
-                      <img
-                        src={result.image}
-                        alt={result.tire}
-                        className="articleImage"
-                      ></img>
+                    <div className="boutonTelechargementDoc">
+                      <Link to="/articlesGrid">
+                        <button className="buttonGreen ">Télécharger</button>
+                      </Link>
                     </div>
                     <div className="articleDetailPara">
                       <p id="articleDetailPara">{result.para1}</p>
                     </div>
-                  </div>
+                    <div className="articleDetailPara">
+                      <p id="articleDetailPara">{result.avantage}</p>
+                    </div>
+                    <div className="boutonTelechargementDoc">
+                      <Link to="/articlesGrid">
+                        <button className="buttonGreen ">Télécharger</button>
+                      </Link>
+                    </div>
 
-                  <div className="boutonTelechargementDoc">
-                    <Link to="/articlesGrid">
-                      <button className="buttonGreen ">Télécharger</button>
-                    </Link>
-                  </div>
-
-                  <p id="articleDetailPara">{result.para1}</p>
-
-                  <div className="boutonTelechargementDoc">
-                    <Link to="/articlesGrid">
-                      <button className="buttonGreen ">Télécharger</button>
-                    </Link>
                   </div>
                 </>
               );
@@ -78,4 +74,4 @@ const ArticleDetail1 = () => {
   );
 };
 
-export default ArticleDetail1;
+export default ArticleDetail;
