--- src/screens/Styles/ArticleDetailcopy.css
@@ -1,151 +1,0 @@
-.detailArticlegGreen {
-  background-color: var(--clr-bg-green);
-  width: 100%;
-  min-height: 100vh;
-  padding: 1rem 0 10rem 0;
-}
-
-.articleDetailWrapper {
-  display: flex;
-  width: clamp(250px, 80vw, 1000px);
-  align-content: center;
-  flex-wrap: wrap;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin: 5rem auto;
-}
-
-.stateCatArticlesDet {
-  padding-top: 2rem;
-}
-
-.stateCatArticlesCat{
-  /* border: 2px solid red; */
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.btnCat{
-  padding: 1rem;
-    border-radius: 0 2rem 0rem 0;
-    width: 20rem;
-    height: 4rem;
-    text-align: center;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    align-content: center;
-    background: var(--clr-orange2);
-    color: var(--clr-white);
-    cursor: pointer;
-    border: solid 5px transparent;
-    box-sizing: border-box;
-}
-
-.gridArrow{
-  width: 30px;
-  height: 30px;
-  transform: rotate(180deg);
-  margin: 0 1rem;
-}
-
-.btnSousCat{
-  padding: 1rem;
-  border-radius: 0 2rem 0rem 0;
-  width: fit-content;
-  min-width: 20rem;
-  height: 4rem;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  align-content: center;
-  background: var(--clr-orange);
-  color: var(--clr-white);
-  cursor: pointer;
-  border: solid 5px transparent;
-  box-sizing: border-box;
-}
-
-.btn-green-form2{
-  margin-left: 2rem;
-}
-
-.articleTitreH2{
-  font-size: 4rem ;
-  text-align: justify;
-  margin-bottom: 1rem;
-}
-
-.articleDetailIntro {
-  width: 100%;
-  font-style: italic;
-  text-align: justify;
-
-}
-
-
-#articleDetailIntro {
-}
-
-.returnButton {
-  display: flex;
-  width: 100%;
-  justify-content: flex-start;
-  padding: 2rem;
-}
-
-.returnButton button {
-  width: 7vw;
-  height: 4vh;
-  background: none;
-  color: var(--clr-white);
-  cursor: pointer;
-}
-
-.returnButton button:hover {
-  color: var(--clr-orange);
-}
-
-.articleDetailWrapper h2 {
-  color: var(--clr-green);
-  font-family: var(--ff-body);
-  font-weight: 600;
-  margin-top: 4rem;
-}
-
-.articleDetailImage {
-  height: auto;
-  width: 100%;
-  margin: 4rem;
-  
-} 
-
-.articleImage{
-  /* border: 10px solid var(--clr-green); */
-  height: 400px;
-  width: 90%;
-  margin: 0 auto;
-  object-fit: cover;
-}
-
-.articleDetailFirst {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-evenly;
-  text-align: justify;
-}
-
-.articleDetailPara {
-  width: 90%;
-  display: flex;
-  margin: 3rem 0;
-}
-
-#articleDetailPara {
-}
